Split out Padding and Size modifiers
